refactor: split some grouped tools out into their own modules (interval and dim_reduction)
