+++ pyaldata/__init__.py
@@ -1,4 +1,6 @@
 from .utils import *
 from .tools import *
+from .interval import *
 from .regression import *
+from .dim_reduction import *
 from .movement_onset_and_peak import *
