+++ pyaldata/tools.py
@@ -2,8 +2,6 @@ import math
 import numpy as np
 import pandas as pd
 from scipy.stats import norm
-from sklearn.decomposition import PCA
-from sklearn.decomposition import FactorAnalysis
 
 from . import utils
 
@@ -550,203 +548,6 @@ def transform_signal(trial_data, signals, transformations, train_trials=None, **
             trial_data = method_dict[trans](trial_data, signal, train_trials, **kwargs)
 
     return trial_data
-
-
-@utils.copy_td
-def restrict_to_interval(trial_data, start_point_name=None, end_point_name=None, rel_start=0, rel_end=0, before=None, after=None, epoch_fun=None, warn_per_trial=False, reset_index=True, ref_field=None):
-    """
-    Restrict time-varying fields to an interval around a time point or between two time points
-
-    trial_data : pd.DataFrame
-        data in trial_data format
-    start_point_name : str, optional
-        name of the time point around which the interval starts
-    end_point_name : str, optional
-        name of the time point around which the interval ends
-        if None, the interval is created around start_point_name
-    rel_start : int, default 0
-        when to start extracting relative to the starting time point
-        replaces the 'before' option
-    rel_end : int, default 0
-        when to stop extracting relative to the ending time point
-        replaces the 'after' option
-    before (deprecated) : int, optional, default None
-        number of time points to extract before the starting time point
-        Please use rel_start instead.
-    after (deprecated): int, optional, default None
-        number of time points to extract after the ending time point
-        Please use rel_end instead.
-    epoch_fun : function, optional
-        function that takes a trial and returns the epoch to extract
-    warn_per_trial : bool, optional, default False
-        give more detailed warnings about indexing in each problematic trial
-    reset_index : bool, optional, default True
-        whether to reset the dataframe index to [0,1,2,...]
-        or keep the original indices of the kept trials
-    ref_field : str (optional)
-        time-varying field to use for identifying the rest
-        if not given, the first field that ends with "spikes" or "rates" is used
-
-    Returns
-    -------
-    data in trial_data format
-    """
-    if before is not None:
-        warnings.warn("'before' is deprecated. Use 'rel_start' instead.")
-        rel_start = -before
-    if after is not None:
-        warnings.warn("'after' is deprecated. Use 'rel_end' instead.")
-        rel_end = after
-
-    assert (start_point_name is None) ^ (epoch_fun is None), "Give either start_point_name or epoch_fun."
-
-    idx_fields = [col for col in trial_data.columns.values if col.startswith("idx")]
-    time_fields = utils.get_time_varying_fields(trial_data, ref_field)
-
-    # generate epoch_fun if the interval is given with time points
-    if start_point_name is not None:
-        epoch_fun = utils.generate_epoch_fun(start_point_name, end_point_name, rel_start, rel_end)
-
-    # check in which trials the indexing works properly
-    kept_trials_mask = np.array([utils._slice_in_trial(trial, epoch_fun(trial), warn_per_trial)
-                                 for (i, trial) in trial_data.iterrows()]).astype(bool)
-    # warn about dropping the problematic trials
-    if np.any(~kept_trials_mask):
-        warnings.warn(f"""Dropping the trials with the following IDs because of invalid time indexing. For more information, try warn_per_trial=True
-
-        {trial_data.trial_id.values[~kept_trials_mask]}""", stacklevel=3)
-
-    # only keep trials in which the indexing works properly
-    trial_data = select_trials(trial_data, kept_trials_mask, reset_index)
-
-    # cut time varying signals
-    for col in time_fields:
-        trial_data[col] = utils.extract_interval_from_signal(trial_data, col, epoch_fun)
-
-    # adjust idx fields
-    def _adjust_field(val, new_T):
-        if isinstance(val, (np.ndarray, list)):
-            return np.array([np.nan
-                             if (idx < 0 or idx > new_T)
-                             else idx
-                             for idx in val])
-        elif ((val < 0) or (val > new_T)):
-            return np.nan
-        else:
-            return val
-
-    new_time_lengths = [arr.shape[0] for arr in trial_data[time_fields[0]]]
-    zero_points = [epoch_fun(trial).start for (i, trial) in trial_data.iterrows()]
-
-    for col in idx_fields:
-        trial_data[col] = [idx - zero_point for (idx, zero_point) in zip(trial_data[col], zero_points)]
-
-        # set indices that are now invalid (i.e. not in the restricted interval) to nan
-        trial_data[col] = [_adjust_field(idx, new_T)
-                           for (idx, new_T) in zip(trial_data[col], new_time_lengths)]
-
-    return trial_data
-
-
-def fit_dim_reduce_model(trial_data, model, signal, train_trials=None, fit_kwargs=None):
-     """
-     Fit a dimensionality reduction model to train_trials
-
-     Parameters
-     ----------
-     trial_data : pd.DataFrame
-         data in trial_data format
-     model : dimensionality reduction model
-         model to fit
-         has to implement a .fit (and .transform) method
-     signal : str
-         signal to fit to
-     train_trials : list of ints (optional)
-         trials to fit the model to
-     fit_kwargs : dict (optional)
-         parameters to pass to model.fit
-
-     Returns
-     -------
-     fitted model
-
-     Example
-     -------
-         from sklearn.decomposition import PCA
-         pca_dims = -5
-         pca = fit_dim_reduce_model(trial_data, PCA(pca_dims), 'M1_rates')
-     """
-     if fit_kwargs is None:
-         fit_kwargs = {}
-
-     model.fit(utils.concat_trials(trial_data, signal, train_trials),
-               **fit_kwargs)
-
-     return model
-
-
-@utils.copy_td
-def apply_dim_reduce_model(trial_data, model, signal, out_fieldname):
-   """
-   Apply a fitted dimensionality reduction model to all trials
-
-   Parameters
-   ----------
-   trial_data : pd.DataFrame
-       data in trial_data format
-   model : dimensionality reduction model
-       fitted model
-       has to implement a .transform method
-   signal : str
-       signal to apply to
-   out_fieldname : str
-       name of the field in which to store the transformed values
-
-   Returns
-   -------
-   trial_data with out_fieldname added
-   """
-   trial_data[out_fieldname] = [model.transform(s) for s in trial_data[signal]]
-
-   return trial_data
-
-
-@utils.copy_td
-def dim_reduce(trial_data, model, signal, out_fieldname, train_trials=None, fit_kwargs=None, return_model=False):
-   """
-   Fit dimensionality reduction model and apply it to all trials
-
-   Parameters
-   ----------
-   trial_data : pd.DataFrame
-       data in trial_data format
-   model : dimensionality reduction model
-       model to fit
-       has to implement a .fit and .transform method
-   signal : str
-       signal to fit and transform
-   out_fieldname : str
-       name of the field in which to store the transformed values
-   train_trials : list of ints (optional)
-       trials to fit the model to
-   fit_kwargs : dict (optional)
-       parameters to pass to model.fit
-   return_model : bool (optional, default False)
-       return the fitted model along with the data
-
-   Returns
-   -------
-   if return_model is False
-       trial_data with the projections added in out_fieldname
-   if return_model is True
-       (trial_data, model)
-   """
-   model = fit_dim_reduce_model(trial_data, model, signal, train_trials, fit_kwargs)
-
-   if return_model:
-       return (apply_dim_reduce_model(trial_data, model, signal, out_fieldname), model)
-   else:
-       return apply_dim_reduce_model(trial_data, model, signal, out_fieldname)
 
 
 def concat_TDs(frames, re_index=True):
