+++ pyaldata/utils.py
@@ -452,151 +452,6 @@ def get_trial_length(trial, ref_field=None):
     return np.size(trial[ref_field], axis=0)
 
 
-def slice_around_index(idx, before, after):
-    """
-    Return a slice around an index
-    Length will be before + after + 1
-
-    Parameters
-    ----------
-    idx : int
-        index around which to create the interval
-    before : int
-        number of time points before time point
-    after : int
-        number of time points after time point
-
-    Returns
-    -------
-    slice(idx-before, idx+after+1)
-    """
-    start = idx - before
-    end = idx + after + 1
-
-    if np.isfinite(start):
-        start = int(start)
-    if np.isfinite(end):
-        end = int(end)
-
-    return slice(start, end)
-
-
-def slice_around_point(trial, point_name, before, after):
-    """
-    Return a slice around a time point in the trial
-    Length will be before + after + 1
-
-    Parameters
-    ----------
-    trial : pd.Series
-        a row from a trial_data dataframe
-        representing a trial
-    point_name : str
-        name of the time point around which to create the interval
-    before : int
-        number of time points before time point
-    after : int
-        number of time points after time point
-
-    Returns
-    -------
-    slice object
-    """
-    start = trial[point_name] - before
-    end = trial[point_name] + after + 1
-
-    if np.isfinite(start):
-        start = int(start)
-    if np.isfinite(end):
-        end = int(end)
-
-    return slice(start, end)
-
-
-def slice_between_points(trial, start_point_name, end_point_name, before, after):
-    """
-    Return a slice that starts before start_point_name and ends after end_point_name
-
-    Parameters
-    ----------
-    trial : pd.Series
-        a row from a trial_data dataframe
-        representing a trial
-    start_point_name : str
-        name of the time point around which the interval starts
-    end_point_name : str
-        name of the time point around which the interval ends
-    before : int
-        number of time points before time point
-    after : int
-        number of time points after time point
-
-    Returns
-    -------
-    slice object
-    """
-    start = trial[start_point_name] - before
-    end = trial[end_point_name] + after + 1
-
-    if np.isfinite(start):
-        start = int(start)
-    if np.isfinite(end):
-        end = int(end)
-
-    return slice(start, end)
-
-
-def generate_epoch_fun(start_point_name, end_point_name=None, rel_start=0, rel_end=0):
-    """
-    Return a function that slices a trial around/between time points
-
-    Parameters
-    ----------
-    start_point_name : str
-        name of the time point around which the interval starts
-    end_point_name : str, optional
-        name of the time point around which the interval ends
-        if None, the interval is created around start_point_name
-    rel_start : int, default 0
-        when to start extracting relative to the starting time point
-        replaces the 'before' option
-    rel_end : int, default 0
-        when to stop extracting relative to the ending time point
-        replaces the 'after' option
-
-    Returns
-    -------
-    epoch_fun : function
-        function that can be used to extract the interval from a trial
-    """
-    if end_point_name is None:
-        epoch_fun = lambda trial: slice_around_point(trial, start_point_name, -rel_start, rel_end)
-    else:
-        epoch_fun = lambda trial: slice_between_points(trial, start_point_name, end_point_name, -rel_start, rel_end)
-
-    return epoch_fun
-
-
-def extract_interval_from_signal(trial_data, signal, epoch_fun):
-    """
-    Extract an interval from a time-varying signal in the dataset
-
-    Parameters
-    ----------
-    trial_data : pd.DataFrame
-        data in trial_data format
-    signal : str
-        name of the signal to extract
-    epoch_fun : function
-        function that takes a trial and returns the epoch to extract
-
-    Returns
-    -------
-    list of the extracted np.arrays
-    """
-    return [trial[signal][epoch_fun(trial), ...] for (i, trial) in trial_data.iterrows()]
-
-
 def remove_cmp_formatting(s):
     """
     Used in read_cmp() to remove formatting in .cmp file
@@ -639,49 +494,6 @@ def read_cmp(file_path):
     df_array = df.apply(pd.to_numeric, errors='ignore')
     
     return df_array
-
-
-def _slice_in_trial(trial, sl, warn=False):
-    """
-    Check if the slice is within the trial's time indices
-
-    Parameters
-    ----------
-    trial : pd.Series
-        trial to check
-    sl : slice
-        slice to check
-    warn : bool, optional, default False
-        whether to warn if the slice is outside
-        the trial's time index
-
-    Returns
-    -------
-    is_inside : bool
-    """
-    T = get_trial_length(trial)
-
-    is_inside = True
-
-    if (sl.start < 0):
-        is_inside = False
-        if warn:
-            warnings.warn(f"Invalid time index on trial with ID {trial.trial_id}. Trying to access index {sl.start} < 0")
-    if (sl.stop > T):
-        is_inside = False
-        if warn:
-            warnings.warn(f"Invalid time index on trial with ID {trial.trial_id}. Trying to access index {sl.stop-1} >= {T}")
-
-    if not np.isfinite(sl.start):
-        is_inside = False
-        if warn:
-            warnings.warn(f"Invalid time index on trial with ID {trial.trial_id}. Starting point is {sl.start}")
-    if not np.isfinite(sl.stop):
-        is_inside = False
-        if warn:
-            warnings.warn(f"Invalid time index on trial with ID {trial.trial_id}. End point is {sl.stop}")
-
-    return is_inside
 
 
 @copy_td
